% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfapply.R
\name{df_apply}
\alias{df_apply}
\title{Apply a function that can be applied to a columns of a data frame (rounding number)}
\usage{
df_apply(.data, .fun, .filter)
}
\arguments{
\item{.data}{should expect a data frame}

\item{.fun}{expect a function that can be applied to a column of a data frame and return a column of a data frame}

\item{.filter}{should expect a function that can be applied to a columns of a data frame and returns a logical (TRUE or FALSE)}
}
\value{
data frame that is a modified version of .data
}
\description{
Apply a function that can be applied to a columns of a data frame (rounding number)
}
\examples{

TestData <- data.frame(A = c(1.092, 300.81, 399.203, 20.4934),
B = c(503.47, 29.89, 293.90, 3.01),
C = c(80.29, 19.867, 39.20, 182.93))

df_apply(TestData, round, is.numeric)
}
